refactor: remove ParametersRepository and StationsRepository interfaces
