+++ src/core/membership/interfaces/index.ts
@@ -1,0 +1,1 @@
+export { UsersRepository } from './users-repository'
